backend/requirements.txt
# Marzban-Manager Backend Dependencies

# Database
psycopg2-binary>=2.9.9

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Validation & Settings
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Security / Encryption
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0

# HTTP Client (for Marzban API calls)
httpx>=0.26.0

# SSH (for node installation)
paramiko>=3.4.0

# Environment
python-dotenv>=1.0.0

# Type hints
typing-extensions>=4.9.0

# Logging
structlog>=24.1.0

# Background tasks
anyio>=4.2.0
